Added post auth

+++ Backend/controllers/postControllers.js
@@ -1,5 +1,6 @@
 import asyncHandler from 'express-async-handler';
 import Post from '../model/postModel.js';
+import User from '../model/userModel.js';
 
 const getSpecificPost = asyncHandler(async (req, res) => {
     const SpecificPost = await Post.findById(req.params.id)
@@ -7,7 +8,7 @@ const getSpecificPost = asyncHandler(async (req, res) => {
 })
 
 const getPost = asyncHandler(async (req, res) => {
-    const post = await Post.find()
+    const post = await Post.find({ user: req.user.id })
     res.status(200).json(post);
 })
 
@@ -18,20 +19,46 @@ const createPost = asyncHandler(async (req, res) => {
     }
     const post = await Post.create({
         text: req.body.text,
+        user: req.user.id,
     })
     res.status(200).json(post);
 })
 
 
 const updatePost = asyncHandler(async (req, res) => {
-    const updatedPost = await Post.findByIdAndUpdate(req.params.id, req.body, { new: true });
+    const post = await Post.findById(req.params.id)
 
-    if (!updatedPost) {
+
+    if (!post) {
         res.status(400);
         throw new Error("Post not found");
     }
 
-    res.status(200).json(updatedPost);
+    const user = await User.findById(req.user.id)
+
+    if (!user) {
+        res.status(400)
+        throw new Error("User not Found")
+
+    }
+
+    if (post.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("Post not authorized");
+
+    }
+    const updatedPost = await Post.findByIdAndUpdate(req.params.id, req.body, { new: true });
+
+    res.status(200).json(updatedPost)
+
+    // const updatedPost = await Post.findByIdAndUpdate(req.params.id, req.body, { new: true });
+
+    // if (!updatedPost) {
+    //     res.status(400);
+    //     throw new Error("Post not found");
+    // }
+
+    // res.status(200).json(updatedPost);
 });
 
 const deletePost = asyncHandler(async (req, res) => {
@@ -41,6 +68,19 @@ const deletePost = asyncHandler(async (req, res) => {
         throw new Error("Post not found")
     }
 
+    const user = await User.findById(req.user.id)
+
+    if (!user) {
+        res.status(400)
+        throw new Error("User not Found")
+
+    }
+
+    if (post.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("Post not authorized");
+
+    }
     await Post.deleteOne({ _id: req.params.id })
 
     res.status(200).json({
